80's Retro style scereen saver
screen rotation and random 256 to 565 colors and random drawings of Triangles.  So this Sketch uses 18334 bytes (56%) of program storage space of the Arduino UNO R3 and the Maximum is 32256 bytes. Global variables use are 257 bytes (12%) of dynamic memory, leaving 1791 bytes for local variables. Maximum is 2048 bytes. The code is A 2nd draft its ugly but if functions.